Dispatch CaseDeleted; handle record deletion in trait
Record deletion moved to trait to resolve sonarqube complexity issue (too many methods in DeleteCase)
